requirements.txt
absl-py==2.1.0
altair==5.2.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
dominate==2.9.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.14.0
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.5.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keras==3.3.3
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.3
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.2
psutil==6.0.0
pure-eval==0.2.2
pyarrow==14.0.2
pycparser==2.22
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.6.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.32.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.30.0
sympy==1.12.1
tbb==2021.12.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.2.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
urllib3==2.1.0
validators==0.22.0
visitor==0.1.3
watchdog==3.0.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
zipp==3.17.0
